+++ trunk/src/utest/srs_utest_config.cpp
@@ -1967,8 +1967,8 @@ VOID TEST(ConfigUnitTest, CheckDefaultValuesVhost)
 
     if (true) {
 	    HELPER_ASSERT_SUCCESS(conf.parse(_MIN_OK_CONF));
-	    EXPECT_EQ(30 * SRS_UTIME_SECONDS, conf.get_dash_fragment(""));
-	    EXPECT_EQ(150 * SRS_UTIME_SECONDS, conf.get_dash_update_period(""));
+	    EXPECT_EQ(10 * SRS_UTIME_SECONDS, conf.get_dash_fragment(""));
+	    EXPECT_EQ(5 * SRS_UTIME_SECONDS, conf.get_dash_update_period(""));
 	    EXPECT_EQ(300 * SRS_UTIME_SECONDS, conf.get_dash_timeshift(""));
 
 	    HELPER_ASSERT_SUCCESS(conf.parse(_MIN_OK_CONF "vhost v{dash{dash_fragment 4;dash_update_period 40;dash_timeshift 70;}}"));
@@ -3865,3 +3865,27 @@ VOID TEST(ConfigMainTest, LogLevelV2)
         EXPECT_EQ(SrsLogLevelWarn, srs_get_log_level(conf.get_log_level_v2()));
     }
 }
+
+VOID TEST(ConfigMainTest, SrtServerTlpktDrop)
+{
+    srs_error_t err;
+
+    if (true) {
+        MockSrsConfig conf;
+        HELPER_ASSERT_SUCCESS(conf.parse(_MIN_OK_CONF));
+        EXPECT_TRUE(conf.get_srto_tlpktdrop());
+    }
+
+    if (true) {
+        MockSrsConfig conf;
+        HELPER_ASSERT_SUCCESS(conf.parse(_MIN_OK_CONF "srt_server{tlpktdrop off;}"));
+        EXPECT_FALSE(conf.get_srto_tlpktdrop());
+    }
+
+    if (true) {
+        MockSrsConfig conf;
+        HELPER_ASSERT_SUCCESS(conf.parse(_MIN_OK_CONF "srt_server{tlpkdrop off;}"));
+        EXPECT_FALSE(conf.get_srto_tlpktdrop());
+    }
+}
+